@@ -14,37 +14,33 @@
     }
   ],
   "require": {
-    "php": "^7.1",
-    "mimmi20/json-class": "~2.0.8",
-    "psr/log": "^1.1.2",
+    "php": "^7.4.3 || ^8.0.0",
+    "ext-mbstring": "*",
+    "psr/log": "^1.1.4",
     "psr/simple-cache": "^1.0.1",
-    "symfony/finder": "^4.4.2",
-    "symfony/yaml": "^4.4.2"
+    "symfony/finder": "^v5.4.2 || ^v6.0.2",
+    "symfony/yaml": "^v5.4.2 || ^v6.0.2"
   },
   "require-dev": {
     "ext-json": "*",
     "ext-pdo": "*",
     "ext-simplexml": "*",
-    "browscap/browscap": "dev-master",
-    "donatj/phpuseragentparser": "dev-master",
-    "ergebnis/composer-normalize": "^2.2.2",
-    "ergebnis/phpstan-rules": "^0.14.2",
-    "friendsofphp/php-cs-fixer": "^2.16.1",
-    "jaybizzle/crawler-detect": "dev-master",
-    "mimmi20/browser-detector": "~5.0.41",
-    "mobiledetect/mobiledetectlib": "dev-master",
-    "pepakriz/phpstan-exception-rules": "^0.10.1",
-    "phpstan/phpstan": "^0.12.5",
-    "phpstan/phpstan-deprecation-rules": "^0.12.2",
-    "phpstan/phpstan-strict-rules": "^0.12.1",
-    "piwik/device-detector": "dev-master",
-    "sinergi/browser-detector": "dev-master",
+    "browscap/browscap": "dev-updates",
+    "donatj/phpuseragentparser": "^v1.5.0",
+    "jaybizzle/crawler-detect": "^v1.2.110",
+    "matomo/device-detector": "dev-master",
+    "mimmi20/browser-detector": "^7.0.0",
+    "mimmi20/coding-standard": "^2.6.0",
+    "mobiledetect/mobiledetectlib": "^2.8.37",
+    "phpstan/extension-installer": "^1.1.0",
+    "phpstan/phpstan": "^1.3.1",
+    "phpstan/phpstan-deprecation-rules": "^1.0.0",
+    "sinergi/browser-detector": "^6.1.4",
     "squizlabs/php_codesniffer": "^3.5.3",
     "ua-parser/uap-core": "dev-master",
-    "whichbrowser/parser": "dev-master",
-    "woothee/woothee-testset": "dev-master",
-    "yzalis/ua-parser": "dev-master",
-    "zsxsoft/php-useragent": "dev-master"
+    "whichbrowser/parser": "^v2.1.2",
+    "woothee/woothee-testset": "^v1.13.0",
+    "zsxsoft/php-useragent": "^1.5"
   },
   "suggest": {
     "ext-http": "to parse headers in the WhichBrowserSource",
@@ -54,20 +50,32 @@
     "browscap/browscap": "To read these useragents",
     "donatj/phpuseragentparser": "To read these useragents",
     "jaybizzle/crawler-detect": "To read these useragents",
+    "matomo/device-detector": "To read these useragents",
     "mobiledetect/mobiledetectlib": "To read these useragents",
     "mongodb/mongodb": "To read these useragents",
-    "piwik/device-detector": "To read these useragents",
     "sinergi/browser-detector": "To read these useragents",
     "symfony/cache": "required for the UapCoreSource",
     "ua-parser/uap-core": "To read these useragents",
     "whichbrowser/parser": "To read these useragents",
     "woothee/woothee-testset": "To read these useragents",
-    "yzalis/ua-parser": "To read these useragents",
     "zsxsoft/php-useragent": "To read these useragents"
+  },
+  "repositories": [
+    {
+      "type": "git",
+      "url": "https://github.com/mimmi20/uap-core"
+    }
+  ],
+  "minimum-stability": "dev",
+  "prefer-stable": true,
+  "autoload": {
+    "psr-4": {
+      "BrowscapHelper\\Source\\": "src/"
+    }
   },
   "config": {
     "platform": {
-      "php": "7.1.3"
+      "php": "7.4.3"
     },
     "preferred-install": {
       "browscap/browscap": "source",
@@ -83,23 +91,5 @@
       "*": "dist"
     },
     "sort-packages": true
-  },
-  "autoload": {
-    "psr-4": {
-      "BrowscapHelper\\Source\\": "src/"
-    }
-  },
-  "autoload-dev": {
-    "psr-4": {
-      "BrowscapHelper\\SourceTest\\": "tests/"
-    }
-  },
-  "repositories": [
-    {
-      "type": "git",
-      "url": "https://github.com/mimmi20/uap-core"
-    }
-  ],
-  "minimum-stability": "dev",
-  "prefer-stable": true
+  }
 }
