@@ -9,53 +9,65 @@
  */
 
 declare(strict_types = 1);
+
 namespace BrowscapHelper\Source;
 
 use BrowscapHelper\Source\Helper\FilePath;
 use BrowscapHelper\Source\Ua\UserAgent;
-use Psr\Log\LoggerInterface;
-use Symfony\Component\Finder\Finder;
+use LogicException;
+use RecursiveDirectoryIterator;
+use RecursiveIteratorIterator;
+use RuntimeException;
+use Symfony\Component\Console\Output\OutputInterface;
 
-final class DirectorySource implements SourceInterface
+use function fclose;
+use function feof;
+use function fgets;
+use function file_exists;
+use function fopen;
+use function mb_strlen;
+use function sprintf;
+use function str_pad;
+use function trim;
+
+use const STR_PAD_RIGHT;
+
+final class DirectorySource implements OutputAwareInterface, SourceInterface
 {
-    use GetUserAgentsTrait;
+    use GetNameTrait;
+    use OutputAwareTrait;
 
-    /**
-     * @var string
-     */
-    private $dir;
+    private const NAME = 'directory-source';
 
-    /**
-     * @var \Psr\Log\LoggerInterface
-     */
-    private $logger;
+    private string $dir;
 
-    /**
-     * @param \Psr\Log\LoggerInterface $logger
-     * @param string                   $dir
-     */
-    public function __construct(LoggerInterface $logger, string $dir)
+    public function __construct(string $dir)
     {
-        $this->logger = $logger;
-        $this->dir    = $dir;
+        $this->dir = $dir;
     }
 
     /**
-     * @return string
+     * @throws void
      */
-    public function getName(): string
+    public function isReady(string $parentMessage): bool
     {
-        return 'directory';
+        if (file_exists($this->dir)) {
+            return true;
+        }
+
+        $this->writeln("\r" . '<error>' . $parentMessage . sprintf('- path %s not found</error>', $this->dir), OutputInterface::VERBOSITY_NORMAL);
+
+        return false;
     }
 
     /**
-     * @throws \LogicException
+     * @return iterable<array<non-empty-string, non-empty-string>>
      *
-     * @return array[]|iterable
+     * @throws RuntimeException
      */
-    public function getHeaders(): iterable
+    public function getHeaders(string $message, int &$messageLength = 0): iterable
     {
-        foreach ($this->loadFromPath() as $line) {
+        foreach ($this->loadFromPath($message, $messageLength) as $line) {
             $ua    = UserAgent::fromUseragent($line);
             $agent = (string) $ua;
 
@@ -68,13 +80,15 @@ final class DirectorySource implements SourceInterface
     }
 
     /**
-     * @throws \LogicException
+     * @return iterable<array<mixed>>
+     * @phpstan-return iterable<array{headers: array<non-empty-string, non-empty-string>, device: array{deviceName: string|null, marketingName: string|null, manufacturer: string|null, brand: string|null, display: array{width: int|null, height: int|null, touch: bool|null, type: string|null, size: float|int|null}, type: string|null, ismobile: bool|null}, client: array{name: string|null, modus: string|null, version: string|null, manufacturer: string|null, bits: int|null, type: string|null, isbot: bool|null}, platform: array{name: string|null, marketingName: string|null, version: string|null, manufacturer: string|null, bits: int|null}, engine: array{name: string|null, version: string|null, manufacturer: string|null}}>
      *
-     * @return array[]|iterable
+     * @throws LogicException
+     * @throws RuntimeException
      */
-    public function getProperties(): iterable
+    public function getProperties(string $message, int &$messageLength = 0): iterable
     {
-        foreach ($this->loadFromPath() as $line) {
+        foreach ($this->loadFromPath($message, $messageLength) as $line) {
             $ua    = UserAgent::fromUseragent($line);
             $agent = (string) $ua;
 
@@ -82,7 +96,8 @@ final class DirectorySource implements SourceInterface
                 continue;
             }
 
-            yield $agent => [
+            yield [
+                'headers' => ['user-agent' => $agent],
                 'device' => [
                     'deviceName' => null,
                     'marketingName' => null,
@@ -95,18 +110,10 @@ final class DirectorySource implements SourceInterface
                         'type' => null,
                         'size' => null,
                     ],
-                    'dualOrientation' => null,
                     'type' => null,
-                    'simCount' => null,
-                    'market' => [
-                        'regions' => null,
-                        'countries' => null,
-                        'vendors' => null,
-                    ],
-                    'connections' => null,
                     'ismobile' => null,
                 ],
-                'browser' => [
+                'client' => [
                     'name' => null,
                     'modus' => null,
                     'version' => null,
@@ -132,51 +139,50 @@ final class DirectorySource implements SourceInterface
     }
 
     /**
-     * @throws \LogicException
+     * @return iterable<array<non-empty-string, non-empty-string>>
      *
-     * @return iterable|string[]
+     * @throws RuntimeException
      */
-    private function loadFromPath(): iterable
+    private function loadFromPath(string $parentMessage, int &$messageLength = 0): iterable
     {
-        if (!file_exists($this->dir)) {
-            $this->logger->warning(sprintf('    path %s not found', $this->dir));
+        $message = $parentMessage . sprintf('- reading path %s', $this->dir);
 
-            return;
+        if (mb_strlen($message) > $messageLength) {
+            $messageLength = mb_strlen($message);
         }
 
-        $this->logger->info(sprintf('    reading path %s', $this->dir));
+        $this->write("\r" . '<info>' . str_pad($message, $messageLength, ' ', STR_PAD_RIGHT) . '</info>', false, OutputInterface::VERBOSITY_VERBOSE);
 
-        $finder = new Finder();
-        $finder->files();
-        $finder->ignoreDotFiles(true);
-        $finder->ignoreVCS(true);
-        $finder->sortByName();
-        $finder->ignoreUnreadableDirs();
-        $finder->in($this->dir);
+        $files = new RecursiveIteratorIterator(new RecursiveDirectoryIterator($this->dir));
 
         $fileHelper = new FilePath();
 
-        foreach ($finder as $file) {
+        foreach ($files as $file) {
             $filepath = $file->getPathname();
 
-            $this->logger->info('    reading file ' . str_pad($filepath, 100, ' ', STR_PAD_RIGHT));
+            $message = $parentMessage . sprintf('- reading file %s', $filepath);
+
+            if (mb_strlen($message) > $messageLength) {
+                $messageLength = mb_strlen($message);
+            }
+
+            $this->write("\r" . '<info>' . str_pad($message, $messageLength, ' ', STR_PAD_RIGHT) . '</info>', false, OutputInterface::VERBOSITY_VERY_VERBOSE);
 
             $fullPath = $fileHelper->getPath($file);
 
             if (null === $fullPath) {
-                $this->logger->error('could not detect path for file "' . $filepath . '"');
-
+                $this->writeln('', OutputInterface::VERBOSITY_VERBOSE);
+                $this->writeln('<error>could not detect path for file "' . $filepath . '"</error>', OutputInterface::VERBOSITY_NORMAL);
                 continue;
             }
 
-            $handle = @fopen($fullPath, 'r');
+            $handle = @fopen($filepath, 'r');
 
             if (false === $handle) {
-                $this->logger->emergency(new \RuntimeException('reading file ' . $filepath . ' caused an error'));
+                $this->writeln('', OutputInterface::VERBOSITY_VERBOSE);
+                $this->writeln('<error>reading file ' . $filepath . ' caused an error</error>', OutputInterface::VERBOSITY_NORMAL);
                 continue;
             }
-
-            $i = 1;
 
             while (!feof($handle)) {
                 $line = fgets($handle, 65535);
@@ -185,11 +191,9 @@ final class DirectorySource implements SourceInterface
                     continue;
                 }
 
-                ++$i;
-
                 $line = trim($line);
 
-                if (empty($line)) {
+                if ('' === $line) {
                     continue;
                 }
 
