@@ -9,54 +9,65 @@
  */
 
 declare(strict_types = 1);
+
 namespace BrowscapHelper\Source;
 
 use BrowscapHelper\Source\Ua\UserAgent;
-use Psr\Log\LoggerInterface;
-use Symfony\Component\Finder\Finder;
+use FilterIterator;
+use Iterator;
+use LogicException;
+use RecursiveDirectoryIterator;
+use RecursiveIteratorIterator;
+use RuntimeException;
+use SplFileInfo;
+use Symfony\Component\Console\Output\OutputInterface;
 use Symfony\Component\Yaml\Yaml;
 
-final class YamlFileSource implements SourceInterface
+use function assert;
+use function file_exists;
+use function is_array;
+use function mb_strlen;
+use function sprintf;
+use function str_pad;
+
+use const STR_PAD_RIGHT;
+
+final class YamlFileSource implements OutputAwareInterface, SourceInterface
 {
-    use GetUserAgentsTrait;
+    use GetNameTrait;
+    use OutputAwareTrait;
 
-    /**
-     * @var string
-     */
-    private $dir;
+    private const NAME = 'yaml-files';
 
-    /**
-     * @var \Psr\Log\LoggerInterface
-     */
-    private $logger;
+    private string $dir;
 
-    /**
-     * @param \Psr\Log\LoggerInterface $logger
-     * @param string                   $dir
-     */
-    public function __construct(LoggerInterface $logger, string $dir)
+    public function __construct(string $dir)
     {
-        $this->logger = $logger;
-        $this->dir    = $dir;
+        $this->dir = $dir;
     }
 
     /**
-     * @return string
+     * @throws void
      */
-    public function getName(): string
+    public function isReady(string $parentMessage): bool
     {
-        return 'yaml-files';
+        if (file_exists($this->dir)) {
+            return true;
+        }
+
+        $this->writeln("\r" . '<error>' . $parentMessage . sprintf('- path %s not found</error>', $this->dir), OutputInterface::VERBOSITY_NORMAL);
+
+        return false;
     }
 
     /**
-     * @throws \LogicException
-     * @throws \RuntimeException
+     * @return iterable<array<non-empty-string, non-empty-string>>
      *
-     * @return array[]|iterable
+     * @throws RuntimeException
      */
-    public function getHeaders(): iterable
+    public function getHeaders(string $message, int &$messageLength = 0): iterable
     {
-        foreach ($this->loadFromPath() as $headers) {
+        foreach ($this->loadFromPath($message, $messageLength) as $headers) {
             $ua    = UserAgent::fromHeaderArray($headers);
             $agent = (string) $ua;
 
@@ -69,14 +80,15 @@ final class YamlFileSource implements SourceInterface
     }
 
     /**
-     * @throws \LogicException
-     * @throws \RuntimeException
+     * @return iterable<array<mixed>>
+     * @phpstan-return iterable<array{headers: array<string, string>, device: array{deviceName: string|null, marketingName: string|null, manufacturer: string|null, brand: string|null, display: array{width: int|null, height: int|null, touch: bool|null, type: string|null, size: float|int|null}, type: string|null, ismobile: bool|null}, client: array{name: string|null, modus: string|null, version: string|null, manufacturer: string|null, bits: int|null, type: string|null, isbot: bool|null}, platform: array{name: string|null, marketingName: string|null, version: string|null, manufacturer: string|null, bits: int|null}, engine: array{name: string|null, version: string|null, manufacturer: string|null}}>
      *
-     * @return array[]|iterable
+     * @throws LogicException
+     * @throws RuntimeException
      */
-    public function getProperties(): iterable
+    public function getProperties(string $message, int &$messageLength = 0): iterable
     {
-        foreach ($this->loadFromPath() as $headers) {
+        foreach ($this->loadFromPath($message, $messageLength) as $headers) {
             $ua    = UserAgent::fromHeaderArray($headers);
             $agent = (string) $ua;
 
@@ -84,7 +96,8 @@ final class YamlFileSource implements SourceInterface
                 continue;
             }
 
-            yield $agent => [
+            yield [
+                'headers' => ['user-agent' => $agent],
                 'device' => [
                     'deviceName' => null,
                     'marketingName' => null,
@@ -97,18 +110,10 @@ final class YamlFileSource implements SourceInterface
                         'type' => null,
                         'size' => null,
                     ],
-                    'dualOrientation' => null,
                     'type' => null,
-                    'simCount' => null,
-                    'market' => [
-                        'regions' => null,
-                        'countries' => null,
-                        'vendors' => null,
-                    ],
-                    'connections' => null,
                     'ismobile' => null,
                 ],
-                'browser' => [
+                'client' => [
                     'name' => null,
                     'modus' => null,
                     'version' => null,
@@ -134,35 +139,54 @@ final class YamlFileSource implements SourceInterface
     }
 
     /**
-     * @throws \LogicException
-     * @throws \RuntimeException
+     * @return iterable<array<non-empty-string, non-empty-string>>
      *
-     * @return array[]|iterable
+     * @throws RuntimeException
      */
-    private function loadFromPath(): iterable
+    private function loadFromPath(string $parentMessage, int &$messageLength = 0): iterable
     {
-        if (!file_exists($this->dir)) {
-            $this->logger->warning(sprintf('    path %s not found', $this->dir));
+        $message = $parentMessage . sprintf('- reading path %s', $this->dir);
 
-            return;
+        if (mb_strlen($message) > $messageLength) {
+            $messageLength = mb_strlen($message);
         }
 
-        $this->logger->info(sprintf('    reading path %s', $this->dir));
+        $this->write("\r" . '<info>' . str_pad($message, $messageLength, ' ', STR_PAD_RIGHT) . '</info>', false, OutputInterface::VERBOSITY_VERBOSE);
 
-        $finder = new Finder();
-        $finder->files();
-        $finder->name('*.yaml');
-        $finder->ignoreDotFiles(true);
-        $finder->ignoreVCS(true);
-        $finder->sortByName();
-        $finder->ignoreUnreadableDirs();
-        $finder->in($this->dir);
+        $iterator = new RecursiveIteratorIterator(new RecursiveDirectoryIterator($this->dir));
+        $files    = new class ($iterator, 'yaml') extends FilterIterator {
+            private string $extension;
 
-        foreach ($finder as $file) {
-            /** @var \Symfony\Component\Finder\SplFileInfo $file */
+            /**
+             * @param Iterator<SplFileInfo> $iterator
+             */
+            public function __construct(Iterator $iterator, string $extension)
+            {
+                parent::__construct($iterator);
+                $this->extension = $extension;
+            }
+
+            public function accept(): bool
+            {
+                $file = $this->getInnerIterator()->current();
+
+                assert($file instanceof SplFileInfo);
+
+                return $file->isFile() && $file->getExtension() === $this->extension;
+            }
+        };
+
+        foreach ($files as $file) {
+            /** @var SplFileInfo $file */
             $filepath = $file->getPathname();
 
-            $this->logger->info('    reading file ' . str_pad($filepath, 100, ' ', STR_PAD_RIGHT));
+            $message = $parentMessage . sprintf('- reading file %s', $filepath);
+
+            if (mb_strlen($message) > $messageLength) {
+                $messageLength = mb_strlen($message);
+            }
+
+            $this->write("\r" . '<info>' . str_pad($message, $messageLength, ' ', STR_PAD_RIGHT) . '</info>', false, OutputInterface::VERBOSITY_VERY_VERBOSE);
 
             $data = Yaml::parse($file->getContents());
 
