@@ -1,24 +1,27 @@
 <?php
 /**
- * This file is part of the browscap-helper-source package.
+ * This file is part of the browscap-helper package.
  *
- * Copyright (c) 2016-2019, Thomas Mueller <mimmi20@live.de>
+ * Copyright (c) 2015-2021, Thomas Mueller <mimmi20@live.de>
  *
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
  */
 
 declare(strict_types = 1);
+
 namespace BrowscapHelper\Source\Helper;
+
+use SplFileInfo;
+
+use function realpath;
 
 final class FilePath
 {
     /**
-     * @param \SplFileInfo $file
-     *
-     * @return string|null
+     * @throws void
      */
-    public function getPath(\SplFileInfo $file): ?string
+    public function getPath(SplFileInfo $file): ?string
     {
         $realpath = realpath($file->getPathname());
 
