@@ -1,63 +1,81 @@
 <?php
 /**
- * This file is part of the browscap-helper-source package.
+ * This file is part of the browscap-helper package.
  *
- * Copyright (c) 2016-2019, Thomas Mueller <mimmi20@live.de>
+ * Copyright (c) 2015-2021, Thomas Mueller <mimmi20@live.de>
  *
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
  */
 
 declare(strict_types = 1);
+
 namespace BrowscapHelper\Source;
 
 use BrowscapHelper\Source\Ua\UserAgent;
-use ExceptionalJSON\DecodeErrorException;
-use JsonClass\Json;
-use Psr\Log\LoggerInterface;
+use Exception;
+use JsonException;
+use LogicException;
+use RuntimeException;
+use SplFileInfo;
+use Symfony\Component\Console\Output\OutputInterface;
 use Symfony\Component\Finder\Finder;
 
-final class JsonFileSource implements SourceInterface
+use function assert;
+use function file_exists;
+use function file_get_contents;
+use function is_array;
+use function is_string;
+use function json_decode;
+use function mb_strlen;
+use function sprintf;
+use function str_pad;
+use function str_replace;
+use function unlink;
+
+use const JSON_THROW_ON_ERROR;
+use const PHP_EOL;
+use const STR_PAD_RIGHT;
+
+final class JsonFileSource implements OutputAwareInterface, SourceInterface
 {
-    use GetUserAgentsTrait;
+    use GetNameTrait;
+    use OutputAwareTrait;
+
+    private const NAME = 'json-files';
+
+    private string $dir;
 
     /**
-     * @var string
+     * @throws void
      */
-    private $dir;
-
-    /**
-     * @var \Psr\Log\LoggerInterface
-     */
-    private $logger;
-
-    /**
-     * @param \Psr\Log\LoggerInterface $logger
-     * @param string                   $dir
-     */
-    public function __construct(LoggerInterface $logger, string $dir)
+    public function __construct(string $dir)
     {
-        $this->logger = $logger;
-        $this->dir    = $dir;
+        $this->dir = $dir;
     }
 
     /**
-     * @return string
+     * @throws void
      */
-    public function getName(): string
+    public function isReady(string $parentMessage): bool
     {
-        return 'json-files';
+        if (file_exists($this->dir)) {
+            return true;
+        }
+
+        $this->writeln("\r" . '<error>' . $parentMessage . sprintf('- path %s not found</error>', $this->dir), OutputInterface::VERBOSITY_NORMAL);
+
+        return false;
     }
 
     /**
-     * @throws \LogicException
-     * @throws \RuntimeException
+     * @return iterable<array<non-empty-string, non-empty-string>>
      *
-     * @return array[]|iterable
+     * @throws RuntimeException
      */
-    public function getHeaders(): iterable
+    public function getHeaders(string $message, int &$messageLength = 0): iterable
     {
-        foreach ($this->loadFromPath() as $headers) {
+        foreach ($this->loadFromPath($message, $messageLength) as $headers) {
             $ua    = UserAgent::fromHeaderArray($headers);
             $agent = (string) $ua;
 
@@ -70,22 +88,24 @@ final class JsonFileSource implements SourceInterface
     }
 
     /**
-     * @throws \LogicException
-     * @throws \RuntimeException
+     * @return iterable<array<mixed>>
+     * @phpstan-return iterable<array{headers: array<string, string>, device: array{deviceName: string|null, marketingName: string|null, manufacturer: string|null, brand: string|null, display: array{width: int|null, height: int|null, touch: bool|null, type: string|null, size: float|int|null}, type: string|null, ismobile: bool|null}, client: array{name: string|null, modus: string|null, version: string|null, manufacturer: string|null, bits: int|null, type: string|null, isbot: bool|null}, platform: array{name: string|null, marketingName: string|null, version: string|null, manufacturer: string|null, bits: int|null}, engine: array{name: string|null, version: string|null, manufacturer: string|null}}>
      *
-     * @return array[]|iterable
+     * @throws LogicException
+     * @throws RuntimeException
      */
-    public function getProperties(): iterable
+    public function getProperties(string $message, int &$messageLength = 0): iterable
     {
-        foreach ($this->loadFromPath() as $headers) {
-            $ua    = UserAgent::fromHeaderArray($headers);
+        foreach ($this->loadFromPath($message, $messageLength) as $line) {
+            $ua    = UserAgent::fromUseragent($line);
             $agent = (string) $ua;
 
             if (empty($agent)) {
                 continue;
             }
 
-            yield $agent => [
+            yield [
+                'headers' => ['user-agent' => $agent],
                 'device' => [
                     'deviceName' => null,
                     'marketingName' => null,
@@ -98,18 +118,10 @@ final class JsonFileSource implements SourceInterface
                         'type' => null,
                         'size' => null,
                     ],
-                    'dualOrientation' => null,
                     'type' => null,
-                    'simCount' => null,
-                    'market' => [
-                        'regions' => null,
-                        'countries' => null,
-                        'vendors' => null,
-                    ],
-                    'connections' => null,
                     'ismobile' => null,
                 ],
-                'browser' => [
+                'client' => [
                     'name' => null,
                     'modus' => null,
                     'version' => null,
@@ -135,20 +147,19 @@ final class JsonFileSource implements SourceInterface
     }
 
     /**
-     * @throws \LogicException
-     * @throws \RuntimeException
+     * @return iterable<array<non-empty-string, non-empty-string>>
      *
-     * @return array[]|iterable
+     * @throws RuntimeException
      */
-    private function loadFromPath(): iterable
+    private function loadFromPath(string $parentMessage, int &$messageLength = 0): iterable
     {
-        if (!file_exists($this->dir)) {
-            $this->logger->warning(sprintf('    path %s not found', $this->dir));
+        $message = $parentMessage . sprintf('- reading path %s', $this->dir);
 
-            return;
+        if (mb_strlen($message) > $messageLength) {
+            $messageLength = mb_strlen($message);
         }
 
-        $this->logger->info(sprintf('    reading path %s', $this->dir));
+        $this->write("\r" . '<info>' . str_pad($message, $messageLength, ' ', STR_PAD_RIGHT) . '</info>', false, OutputInterface::VERBOSITY_VERBOSE);
 
         $finder = new Finder();
         $finder->files();
@@ -160,19 +171,33 @@ final class JsonFileSource implements SourceInterface
         $finder->in($this->dir);
 
         foreach ($finder as $file) {
-            /** @var \Symfony\Component\Finder\SplFileInfo $file */
-            $filepath = $file->getPathname();
+            /** @var SplFileInfo $file */
+            $pathName = $file->getPathname();
+            $filepath = str_replace('\\', '/', $pathName);
+            assert(is_string($filepath));
 
-            $this->logger->info('    reading file ' . str_pad($filepath, 100, ' ', STR_PAD_RIGHT));
+            $message = $parentMessage . sprintf('- reading file %s', $filepath);
+
+            if (mb_strlen($message) > $messageLength) {
+                $messageLength = mb_strlen($message);
+            }
+
+            $this->write("\r" . '<info>' . str_pad($message, $messageLength, ' ', STR_PAD_RIGHT) . '</info>', false, OutputInterface::VERBOSITY_VERY_VERBOSE);
+
+            $content = file_get_contents($filepath);
+
+            if ('' === $content || PHP_EOL === $content) {
+                unlink($filepath);
+
+                continue;
+            }
 
             try {
-                $data = (new Json())->decode(
-                    $file->getContents(),
-                    true
-                );
-            } catch (DecodeErrorException $e) {
-                $this->logger->error(
-                    new \Exception(sprintf('file %s contains invalid json.', $file->getPathname()), 0, $e)
+                $data = json_decode($content, true, 512, JSON_THROW_ON_ERROR);
+            } catch (JsonException $e) {
+                $this->writeln('', OutputInterface::VERBOSITY_VERBOSE);
+                $this->writeln(
+                    '<error>' . (new Exception(sprintf('file %s contains invalid json.', $filepath), 0, $e)) . '</error>'
                 );
                 continue;
             }
