@@ -1,8 +1,8 @@
 <?php
 /**
- * This file is part of the browscap-helper-source package.
+ * This file is part of the browscap-helper package.
  *
- * Copyright (c) 2016-2019, Thomas Mueller <mimmi20@live.de>
+ * Copyright (c) 2015-2021, Thomas Mueller <mimmi20@live.de>
  *
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
@@ -13,37 +13,38 @@ declare(strict_types = 1);
 namespace BrowscapHelper\Source;
 
 use BrowscapHelper\Source\Ua\UserAgent;
-use Exception;
 use FilterIterator;
 use Iterator;
-use JsonException;
 use LogicException;
 use RecursiveDirectoryIterator;
 use RecursiveIteratorIterator;
 use RuntimeException;
 use SplFileInfo;
 use Symfony\Component\Console\Output\OutputInterface;
+use Symfony\Component\Yaml\Yaml;
 
+use function array_map;
 use function assert;
+use function explode;
 use function file_exists;
+use function implode;
 use function is_array;
-use function json_decode;
+use function is_string;
 use function mb_strlen;
 use function sprintf;
 use function str_pad;
+use function str_replace;
 use function trim;
 
-use const JSON_THROW_ON_ERROR;
-use const PHP_EOL;
 use const STR_PAD_RIGHT;
 
-final class DonatjSource implements OutputAwareInterface, SourceInterface
+final class MatomoSource implements OutputAwareInterface, SourceInterface
 {
     use GetNameTrait;
     use OutputAwareTrait;
 
-    private const NAME = 'donatj/phpuseragentparser';
-    private const PATH = 'vendor/donatj/phpuseragentparser/Tests';
+    private const NAME = 'matomo/device-detector';
+    private const PATH = 'vendor/matomo/device-detector/regexes';
 
     /**
      * @throws void
@@ -60,14 +61,18 @@ final class DonatjSource implements OutputAwareInterface, SourceInterface
     }
 
     /**
-     * @return iterable<array<string, string>>
+     * @return iterable<array<non-empty-string, non-empty-string>>
      *
      * @throws RuntimeException
      */
     public function getHeaders(string $message, int &$messageLength = 0): iterable
     {
-        foreach ($this->loadFromPath($message, $messageLength) as $test => $data) {
-            $ua    = UserAgent::fromUseragent(trim($test));
+        foreach ($this->loadFromPath($message, $messageLength) as $row) {
+            $ua    = explode("\n", $row['user_agent']);
+            $ua    = array_map('trim', $ua);
+            $agent = trim(implode(' ', $ua));
+
+            $ua    = UserAgent::fromUseragent($agent);
             $agent = (string) $ua;
 
             if (empty($agent)) {
@@ -87,8 +92,8 @@ final class DonatjSource implements OutputAwareInterface, SourceInterface
      */
     public function getProperties(string $message, int &$messageLength = 0): iterable
     {
-        foreach ($this->loadFromPath($message, $messageLength) as $test => $data) {
-            $ua    = UserAgent::fromUseragent(trim($test));
+        foreach ($this->loadFromPath($message, $messageLength) as $line) {
+            $ua    = UserAgent::fromUseragent($line);
             $agent = (string) $ua;
 
             if (empty($agent)) {
@@ -113,16 +118,16 @@ final class DonatjSource implements OutputAwareInterface, SourceInterface
                     'ismobile' => null,
                 ],
                 'client' => [
-                    'name' => $data['browser'],
+                    'name' => null,
                     'modus' => null,
-                    'version' => $data['version'],
+                    'version' => null,
                     'manufacturer' => null,
                     'bits' => null,
                     'type' => null,
                     'isbot' => null,
                 ],
                 'platform' => [
-                    'name' => $data['platform'],
+                    'name' => null,
                     'marketingName' => null,
                     'version' => null,
                     'manufacturer' => null,
@@ -138,7 +143,7 @@ final class DonatjSource implements OutputAwareInterface, SourceInterface
     }
 
     /**
-     * @return iterable<array<string, string>>
+     * @return array<string, string>|iterable
      *
      * @throws RuntimeException
      */
@@ -153,7 +158,7 @@ final class DonatjSource implements OutputAwareInterface, SourceInterface
         $this->write("\r" . '<info>' . str_pad($message, $messageLength, ' ', STR_PAD_RIGHT) . '</info>', false, OutputInterface::VERBOSITY_VERBOSE);
 
         $iterator = new RecursiveIteratorIterator(new RecursiveDirectoryIterator(self::PATH));
-        $files    = new class ($iterator, 'json') extends FilterIterator {
+        $files    = new class ($iterator, 'yml') extends FilterIterator {
             private string $extension;
 
             /**
@@ -176,8 +181,9 @@ final class DonatjSource implements OutputAwareInterface, SourceInterface
         };
 
         foreach ($files as $file) {
-            /** @var SplFileInfo $file */
-            $filepath = $file->getPathname();
+            $pathName = $file->getPathname();
+            $filepath = str_replace('\\', '/', $pathName);
+            assert(is_string($filepath));
 
             $message = $parentMessage . sprintf('- reading file %s', $filepath);
 
@@ -187,31 +193,18 @@ final class DonatjSource implements OutputAwareInterface, SourceInterface
 
             $this->write("\r" . '<info>' . str_pad($message, $messageLength, ' ', STR_PAD_RIGHT) . '</info>', false, OutputInterface::VERBOSITY_VERY_VERBOSE);
 
-            $content = $file->getContents();
+            $data = Yaml::parseFile($filepath);
 
-            if ('' === $content || PHP_EOL === $content) {
+            if (!is_array($data)) {
                 continue;
             }
 
-            try {
-                $provider = json_decode(
-                    $content,
-                    true,
-                    512,
-                    JSON_THROW_ON_ERROR
-                );
-            } catch (JsonException $e) {
-                $this->logger->critical(new Exception('    parsing file content [' . $filepath . '] failed', 0, $e));
+            foreach ($data as $row) {
+                if (empty($row['user_agent'])) {
+                    continue;
+                }
 
-                continue;
-            }
-
-            if (!is_array($provider)) {
-                continue;
-            }
-
-            foreach ($provider as $test => $data) {
-                yield $test => $data;
+                yield $row;
             }
         }
     }
