@@ -9,48 +9,63 @@
  */
 
 declare(strict_types = 1);
+
 namespace BrowscapHelper\Source;
 
 use BrowscapHelper\Source\Ua\UserAgent;
-use ExceptionalJSON\DecodeErrorException;
-use JsonClass\Json;
-use Psr\Log\LoggerInterface;
-use Symfony\Component\Finder\Finder;
+use Exception;
+use FilterIterator;
+use Iterator;
+use JsonException;
+use LogicException;
+use RecursiveDirectoryIterator;
+use RecursiveIteratorIterator;
+use RuntimeException;
+use SplFileInfo;
+use Symfony\Component\Console\Output\OutputInterface;
 
-final class UaParserJsSource implements SourceInterface
+use function assert;
+use function file_exists;
+use function is_array;
+use function json_decode;
+use function mb_strlen;
+use function sprintf;
+use function str_pad;
+use function trim;
+
+use const JSON_THROW_ON_ERROR;
+use const STR_PAD_RIGHT;
+
+final class UaParserJsSource implements OutputAwareInterface, SourceInterface
 {
-    use GetUserAgentsTrait;
+    use GetNameTrait;
+    use OutputAwareTrait;
+
+    private const NAME = 'ua-parser-js';
+    private const PATH = 'node_modules/ua-parser-js/test';
 
     /**
-     * @var \Psr\Log\LoggerInterface
+     * @throws void
      */
-    private $logger;
-
-    /**
-     * @param \Psr\Log\LoggerInterface $logger
-     */
-    public function __construct(LoggerInterface $logger)
+    public function isReady(string $parentMessage): bool
     {
-        $this->logger = $logger;
+        if (file_exists(self::PATH)) {
+            return true;
+        }
+
+        $this->writeln("\r" . '<error>' . $parentMessage . sprintf('- path %s not found</error>', self::PATH), OutputInterface::VERBOSITY_NORMAL);
+
+        return false;
     }
 
     /**
-     * @return string
-     */
-    public function getName(): string
-    {
-        return 'ua-parser-js';
-    }
-
-    /**
-     * @throws \LogicException
-     * @throws \RuntimeException
+     * @return iterable<array<string, string>>
      *
-     * @return array[]|iterable
+     * @throws RuntimeException
      */
-    public function getHeaders(): iterable
+    public function getHeaders(string $message, int &$messageLength = 0): iterable
     {
-        foreach ($this->loadFromPath() as $providerName => $data) {
+        foreach ($this->loadFromPath($message, $messageLength) as $providerName => $data) {
             $agent = trim($data['ua']);
 
             $ua    = UserAgent::fromUseragent($agent);
@@ -65,15 +80,17 @@ final class UaParserJsSource implements SourceInterface
     }
 
     /**
-     * @throws \LogicException
-     * @throws \RuntimeException
+     * @return iterable<array<mixed>>
+     * @phpstan-return iterable<array{headers: array<string, string>, device: array{deviceName: string|null, marketingName: string|null, manufacturer: string|null, brand: string|null, display: array{width: int|null, height: int|null, touch: bool|null, type: string|null, size: float|int|null}, type: string|null, ismobile: bool|null}, client: array{name: string|null, modus: string|null, version: string|null, manufacturer: string|null, bits: int|null, type: string|null, isbot: bool|null}, platform: array{name: string|null, marketingName: string|null, version: string|null, manufacturer: string|null, bits: int|null}, engine: array{name: string|null, version: string|null, manufacturer: string|null}}>
      *
-     * @return array[]|iterable
+     * @throws LogicException
+     * @throws RuntimeException
      */
-    public function getProperties(): iterable
+    public function getProperties(string $message, int &$messageLength = 0): iterable
     {
         $agents = [];
         $base   = [
+            'headers' => ['user-agent' => $agent],
             'device' => [
                 'deviceName' => null,
                 'marketingName' => null,
@@ -86,18 +103,10 @@ final class UaParserJsSource implements SourceInterface
                     'type' => null,
                     'size' => null,
                 ],
-                'dualOrientation' => null,
                 'type' => null,
-                'simCount' => null,
-                'market' => [
-                    'regions' => null,
-                    'countries' => null,
-                    'vendors' => null,
-                ],
-                'connections' => null,
                 'ismobile' => null,
             ],
-            'browser' => [
+            'client' => [
                 'name' => null,
                 'modus' => null,
                 'version' => null,
@@ -120,7 +129,7 @@ final class UaParserJsSource implements SourceInterface
             ],
         ];
 
-        foreach ($this->loadFromPath() as $providerName => $data) {
+        foreach ($this->loadFromPath($message, $messageLength) as $providerName => $data) {
             $agent = trim($data['ua']);
 
             if (!isset($agents[$agent])) {
@@ -163,46 +172,65 @@ final class UaParserJsSource implements SourceInterface
     }
 
     /**
-     * @throws \LogicException
-     * @throws \RuntimeException
+     * @return iterable<array<string, string>>
      *
-     * @return array[]|iterable
+     * @throws RuntimeException
      */
-    private function loadFromPath(): iterable
+    private function loadFromPath(string $parentMessage, int &$messageLength = 0): iterable
     {
-        $path = 'node_modules/ua-parser-js/test';
+        $message = $parentMessage . sprintf('- reading path %s', self::PATH);
 
-        if (!file_exists($path)) {
-            $this->logger->warning(sprintf('    path %s not found', $path));
-
-            return;
+        if (mb_strlen($message) > $messageLength) {
+            $messageLength = mb_strlen($message);
         }
 
-        $this->logger->info(sprintf('    reading path %s', $path));
+        $this->write("\r" . '<info>' . str_pad($message, $messageLength, ' ', STR_PAD_RIGHT) . '</info>', false, OutputInterface::VERBOSITY_VERBOSE);
 
-        $finder = new Finder();
-        $finder->files();
-        $finder->name('*.json');
-        $finder->ignoreDotFiles(true);
-        $finder->ignoreVCS(true);
-        $finder->sortByName();
-        $finder->ignoreUnreadableDirs();
-        $finder->in($path);
+        $iterator = new RecursiveIteratorIterator(new RecursiveDirectoryIterator(self::PATH));
+        $files    = new class ($iterator, 'json') extends FilterIterator {
+            private string $extension;
 
-        foreach ($finder as $file) {
-            /** @var \Symfony\Component\Finder\SplFileInfo $file */
+            /**
+             * @param Iterator<SplFileInfo> $iterator
+             */
+            public function __construct(Iterator $iterator, string $extension)
+            {
+                parent::__construct($iterator);
+                $this->extension = $extension;
+            }
+
+            public function accept(): bool
+            {
+                $file = $this->getInnerIterator()->current();
+
+                assert($file instanceof SplFileInfo);
+
+                return $file->isFile() && $file->getExtension() === $this->extension;
+            }
+        };
+
+        foreach ($files as $file) {
+            /** @var SplFileInfo $file */
             $filepath = $file->getPathname();
 
-            $this->logger->info('    reading file ' . str_pad($filepath, 100, ' ', STR_PAD_RIGHT));
+            $message = $parentMessage . sprintf('- reading file %s', $filepath);
+
+            if (mb_strlen($message) > $messageLength) {
+                $messageLength = mb_strlen($message);
+            }
+
+            $this->write("\r" . '<info>' . str_pad($message, $messageLength, ' ', STR_PAD_RIGHT) . '</info>', false, OutputInterface::VERBOSITY_VERY_VERBOSE);
 
             try {
-                $provider = (new Json())->decode(
+                $provider = json_decode(
                     $file->getContents(),
-                    true
+                    true,
+                    512,
+                    JSON_THROW_ON_ERROR
                 );
-            } catch (DecodeErrorException $e) {
+            } catch (JsonException $e) {
                 $this->logger->error(
-                    new \Exception(sprintf('file %s contains invalid json.', $file->getPathname()), 0, $e)
+                    new Exception(sprintf('file %s contains invalid json.', $file->getPathname()), 0, $e)
                 );
                 continue;
             }
