@@ -1,56 +1,71 @@
 <?php
 /**
- * This file is part of the browscap-helper-source package.
+ * This file is part of the browscap-helper package.
  *
- * Copyright (c) 2016-2019, Thomas Mueller <mimmi20@live.de>
+ * Copyright (c) 2015-2021, Thomas Mueller <mimmi20@live.de>
  *
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
  */
 
 declare(strict_types = 1);
+
 namespace BrowscapHelper\Source;
 
 use BrowscapHelper\Source\Ua\UserAgent;
-use http\Header;
-use Psr\Log\LoggerInterface;
-use Symfony\Component\Finder\Finder;
+use FilterIterator;
+use Iterator;
+use RecursiveDirectoryIterator;
+use RecursiveIteratorIterator;
+use RuntimeException;
+use SplFileInfo;
+use Symfony\Component\Console\Output\OutputInterface;
 use Symfony\Component\Yaml\Yaml;
 
-final class WhichBrowserSource implements SourceInterface
+use function array_key_exists;
+use function assert;
+use function file_exists;
+use function is_array;
+use function is_string;
+use function mb_strlen;
+use function mb_strpos;
+use function mb_strtolower;
+use function sprintf;
+use function str_pad;
+use function str_replace;
+
+use const STR_PAD_RIGHT;
+
+final class WhichBrowserSource implements OutputAwareInterface, SourceInterface
 {
-    use GetUserAgentsTrait;
+    use GetNameTrait;
+    use OutputAwareTrait;
+
+    private const NAME = 'whichbrowser/parser';
+    private const PATH = 'vendor/whichbrowser/parser/tests/data';
 
     /**
-     * @var \Psr\Log\LoggerInterface
+     * @throws void
      */
-    private $logger;
-
-    /**
-     * @param \Psr\Log\LoggerInterface $logger
-     */
-    public function __construct(LoggerInterface $logger)
+    public function isReady(string $parentMessage): bool
     {
-        $this->logger = $logger;
+        if (file_exists(self::PATH)) {
+            return true;
+        }
+
+        $this->writeln("\r" . '<error>' . $parentMessage . sprintf('- path %s not found</error>', self::PATH), OutputInterface::VERBOSITY_NORMAL);
+
+        return false;
     }
 
     /**
-     * @return string
-     */
-    public function getName(): string
-    {
-        return 'whichbrowser/parser';
-    }
-
-    /**
-     * @throws \LogicException
-     * @throws \RuntimeException
+     * @return iterable<array<non-empty-string, non-empty-string>>
      *
-     * @return array[]|iterable
+     * @throws RuntimeException
      */
-    public function getHeaders(): iterable
+    public function getHeaders(string $message, int &$messageLength = 0): iterable
     {
-        foreach ($this->loadFromPath() as $row) {
+        foreach ($this->loadFromPath($message, $messageLength) as $row) {
             $lowerHeaders = [];
 
             foreach ($this->getHeadersFromRow($row) as $header => $value) {
@@ -69,110 +84,57 @@ final class WhichBrowserSource implements SourceInterface
     }
 
     /**
-     * @throws \LogicException
-     * @throws \RuntimeException
+     * @return array<string, array<string, string>|string>|iterable
      *
-     * @return array[]|iterable
+     * @throws RuntimeException
      */
-    public function getProperties(): iterable
+    private function loadFromPath(string $parentMessage, int &$messageLength = 0): iterable
     {
-        foreach ($this->loadFromPath() as $row) {
-            $lowerHeaders = [];
+        $message = $parentMessage . sprintf('- reading path %s', self::PATH);
 
-            foreach ($this->getHeadersFromRow($row) as $header => $value) {
-                $lowerHeaders[mb_strtolower((string) $header)] = $value;
-            }
-
-            $ua    = UserAgent::fromHeaderArray($lowerHeaders);
-            $agent = (string) $ua;
-
-            if (empty($agent)) {
-                continue;
-            }
-
-            yield $agent => [
-                'device' => [
-                    'deviceName' => $row['device']['model'] ?? null,
-                    'marketingName' => null,
-                    'manufacturer' => null,
-                    'brand' => $row['device']['manufacturer'] ?? null,
-                    'display' => [
-                        'width' => null,
-                        'height' => null,
-                        'touch' => null,
-                        'type' => null,
-                        'size' => null,
-                    ],
-                    'dualOrientation' => null,
-                    'type' => $row['device']['type'] ?? null,
-                    'simCount' => null,
-                    'market' => [
-                        'regions' => null,
-                        'countries' => null,
-                        'vendors' => null,
-                    ],
-                    'connections' => null,
-                    'ismobile' => $this->isMobile($row) ? true : false,
-                ],
-                'browser' => [
-                    'name' => $row['browser']['name'] ?? null,
-                    'modus' => null,
-                    'version' => (!empty($row['browser']['version']) ? is_array($row['browser']['version']) ? $row['browser']['version']['value'] : $row['browser']['version'] : null),
-                    'manufacturer' => null,
-                    'bits' => null,
-                    'type' => null,
-                    'isbot' => null,
-                ],
-                'platform' => [
-                    'name' => $row['os']['name'] ?? null,
-                    'marketingName' => null,
-                    'version' => (!empty($row['os']['version']) ? is_array($row['os']['version']) ? $row['os']['version']['value'] : $row['os']['version'] : null),
-                    'manufacturer' => null,
-                    'bits' => null,
-                ],
-                'engine' => [
-                    'name' => $row['engine']['name'] ?? null,
-                    'version' => $row['engine']['version'] ?? null,
-                    'manufacturer' => null,
-                ],
-            ];
-        }
-    }
-
-    /**
-     * @throws \LogicException
-     * @throws \RuntimeException
-     *
-     * @return array[]|iterable
-     */
-    private function loadFromPath(): iterable
-    {
-        $path = 'vendor/whichbrowser/parser/tests/data';
-
-        if (!file_exists($path)) {
-            $this->logger->warning(sprintf('    path %s not found', $path));
-
-            return;
+        if (mb_strlen($message) > $messageLength) {
+            $messageLength = mb_strlen($message);
         }
 
-        $this->logger->info(sprintf('    reading path %s', $path));
+        $this->write("\r" . '<info>' . str_pad($message, $messageLength, ' ', STR_PAD_RIGHT) . '</info>', false, OutputInterface::VERBOSITY_VERBOSE);
 
-        $finder = new Finder();
-        $finder->files();
-        $finder->name('*.yaml');
-        $finder->ignoreDotFiles(true);
-        $finder->ignoreVCS(true);
-        $finder->sortByName();
-        $finder->ignoreUnreadableDirs();
-        $finder->in($path);
+        $iterator = new RecursiveIteratorIterator(new RecursiveDirectoryIterator(self::PATH));
+        $files    = new class ($iterator, 'yaml') extends FilterIterator {
+            private string $extension;
 
-        foreach ($finder as $file) {
-            /** @var \Symfony\Component\Finder\SplFileInfo $file */
-            $filepath = $file->getPathname();
+            /**
+             * @param Iterator<SplFileInfo> $iterator
+             */
+            public function __construct(Iterator $iterator, string $extension)
+            {
+                parent::__construct($iterator);
+                $this->extension = $extension;
+            }
 
-            $this->logger->info('    reading file ' . str_pad($filepath, 100, ' ', STR_PAD_RIGHT));
+            public function accept(): bool
+            {
+                $file = $this->getInnerIterator()->current();
 
-            $data = Yaml::parse($file->getContents());
+                assert($file instanceof SplFileInfo);
+
+                return $file->isFile() && $file->getExtension() === $this->extension;
+            }
+        };
+
+        foreach ($files as $file) {
+            $pathName = $file->getPathname();
+            $filepath = str_replace('\\', '/', $pathName);
+            assert(is_string($filepath));
+
+            $message = $parentMessage . sprintf('- reading file %s', $filepath);
+
+            if (mb_strlen($message) > $messageLength) {
+                $messageLength = mb_strlen($message);
+            }
+
+            $this->write("\r" . '<info>' . str_pad($message, $messageLength, ' ', STR_PAD_RIGHT) . '</info>', false, OutputInterface::VERBOSITY_VERY_VERBOSE);
+
+            $data = Yaml::parseFile($filepath);
 
             if (!is_array($data)) {
                 continue;
@@ -185,62 +147,26 @@ final class WhichBrowserSource implements SourceInterface
     }
 
     /**
-     * @param array $row
+     * @param array<string, array<string, string>|string> $row
      *
-     * @return array
+     * @return array<string, string>
+     *
+     * @throws void
      */
     private function getHeadersFromRow(array $row): array
     {
-        $headers = [];
-
         if (array_key_exists('headers', $row)) {
             if (is_array($row['headers'])) {
                 return $row['headers'];
             }
 
-            if (class_exists(Header::class)) {
-                // pecl_http versions 2.x/3.x
-                $headers = Header::parse($row['headers']);
-            } elseif (function_exists('\http_parse_headers')) {
-                // pecl_http version 1.x
-                $headers = \http_parse_headers($row['headers']);
-            } elseif (0 === mb_strpos($row['headers'], 'User-Agent: ')) {
-                $headers = ['user-agent' => str_replace('User-Agent: ', '', $row['headers'])];
-            } else {
-                return [];
+            if (is_string($row['headers']) && 0 === mb_strpos($row['headers'], 'User-Agent: ')) {
+                return ['user-agent' => str_replace('User-Agent: ', '', $row['headers'])];
             }
-        }
 
-        if (is_array($headers)) {
-            return $headers;
+            return [];
         }
 
         return [];
-    }
-
-    /**
-     * @param array $data
-     *
-     * @return bool
-     */
-    private function isMobile(array $data): bool
-    {
-        if (!isset($data['device']['type'])) {
-            return false;
-        }
-
-        $mobileTypes = ['mobile', 'tablet', 'ereader', 'media', 'watch', 'camera'];
-
-        if (in_array($data['device']['type'], $mobileTypes, true)) {
-            return true;
-        }
-
-        if ('gaming' === $data['device']['type']) {
-            if (isset($data['device']['subtype']) && 'portable' === $data['device']['subtype']) {
-                return true;
-            }
-        }
-
-        return false;
     }
 }
