@@ -9,45 +9,65 @@
  */
 
 declare(strict_types = 1);
+
 namespace BrowscapHelper\Source;
 
 use BrowscapHelper\Source\Ua\UserAgent;
-use Psr\Log\LoggerInterface;
-use Symfony\Component\Finder\Finder;
+use FilterIterator;
+use Iterator;
+use LogicException;
+use RecursiveDirectoryIterator;
+use RecursiveIteratorIterator;
+use RuntimeException;
+use SplFileInfo;
+use SplFileObject;
+use Symfony\Component\Console\Output\OutputInterface;
 
-final class ZsxsoftSource implements SourceInterface
+use function array_unique;
+use function assert;
+use function count;
+use function file_exists;
+use function mb_strlen;
+use function mb_strpos;
+use function preg_match;
+use function sprintf;
+use function str_pad;
+use function str_replace;
+use function trim;
+use function usort;
+
+use const STR_PAD_RIGHT;
+
+final class ZsxsoftSource implements OutputAwareInterface, SourceInterface
 {
-    use GetUserAgentsTrait;
+    use GetNameTrait;
+    use OutputAwareTrait;
+
+    private const NAME = 'zsxsoft/php-useragent';
+    private const PATH = 'vendor/zsxsoft/php-useragent/lib';
 
     /**
-     * @var \Psr\Log\LoggerInterface
+     * @throws void
      */
-    private $logger;
-
-    /**
-     * @param \Psr\Log\LoggerInterface $logger
-     */
-    public function __construct(LoggerInterface $logger)
+    public function isReady(string $parentMessage): bool
     {
-        $this->logger = $logger;
+        if (file_exists(self::PATH)) {
+            return true;
+        }
+
+        $this->writeln("\r" . '<error>' . $parentMessage . sprintf('- path %s not found</error>', self::PATH), OutputInterface::VERBOSITY_NORMAL);
+
+        return false;
     }
 
     /**
-     * @return string
-     */
-    public function getName(): string
-    {
-        return 'zsxsoft/php-useragent';
-    }
-
-    /**
-     * @throws \LogicException
+     * @return iterable<array<string, string>>
      *
-     * @return array[]|iterable
+     * @throws RuntimeException
      */
-    public function getHeaders(): iterable
+    public function getHeaders(string $message, int &$messageLength = 0): iterable
     {
-        foreach ($this->loadFromPath() as $data) {
+        foreach ($this->loadFromPath($message, $messageLength) as $data) {
             $agent = trim($data[0][0]);
 
             $ua    = UserAgent::fromUseragent($agent);
@@ -62,16 +82,17 @@ final class ZsxsoftSource implements SourceInterface
     }
 
     /**
-     * @throws \LogicException
-     * @throws \RuntimeException
+     * @return iterable<array<mixed>>
+     * @phpstan-return iterable<array{headers: array<string, string>, device: array{deviceName: string|null, marketingName: string|null, manufacturer: string|null, brand: string|null, display: array{width: int|null, height: int|null, touch: bool|null, type: string|null, size: float|int|null}, type: string|null, ismobile: bool|null}, client: array{name: string|null, modus: string|null, version: string|null, manufacturer: string|null, bits: int|null, type: string|null, isbot: bool|null}, platform: array{name: string|null, marketingName: string|null, version: string|null, manufacturer: string|null, bits: int|null}, engine: array{name: string|null, version: string|null, manufacturer: string|null}}>
      *
-     * @return array[]|iterable
+     * @throws LogicException
+     * @throws RuntimeException
      */
-    public function getProperties(): iterable
+    public function getProperties(string $message, int &$messageLength = 0): iterable
     {
         $brands = $this->getBrands();
 
-        foreach ($this->loadFromPath() as $data) {
+        foreach ($this->loadFromPath($message, $messageLength) as $data) {
             $agent = trim($data[0][0]);
 
             $ua    = UserAgent::fromUseragent($agent);
@@ -93,12 +114,13 @@ final class ZsxsoftSource implements SourceInterface
                 $brand = '';
             }
 
-            yield $agent => [
+            yield [
+                'headers' => ['user-agent' => $agent],
                 'device' => [
                     'deviceName' => $model,
                     'marketingName' => null,
                     'manufacturer' => null,
-                    'brand' => $brand ?? '',
+                    'brand' => $brand ?? null,
                     'display' => [
                         'width' => null,
                         'height' => null,
@@ -106,18 +128,10 @@ final class ZsxsoftSource implements SourceInterface
                         'type' => null,
                         'size' => null,
                     ],
-                    'dualOrientation' => null,
                     'type' => null,
-                    'simCount' => null,
-                    'market' => [
-                        'regions' => null,
-                        'countries' => null,
-                        'vendors' => null,
-                    ],
-                    'connections' => null,
                     'ismobile' => null,
                 ],
-                'browser' => [
+                'client' => [
                     'name' => $data[1][2],
                     'modus' => null,
                     'version' => $data[1][3],
@@ -143,36 +157,54 @@ final class ZsxsoftSource implements SourceInterface
     }
 
     /**
-     * @throws \LogicException
+     * @return iterable<array<string, string>>
      *
-     * @return array[]|iterable
+     * @throws RuntimeException
      */
-    private function loadFromPath(): iterable
+    private function loadFromPath(string $parentMessage, int &$messageLength = 0): iterable
     {
-        $path = 'vendor/zsxsoft/php-useragent/tests';
+        $message = $parentMessage . sprintf('- reading path %s', self::PATH);
 
-        if (!file_exists($path)) {
-            $this->logger->warning(sprintf('    path %s not found', $path));
-
-            return;
+        if (mb_strlen($message) > $messageLength) {
+            $messageLength = mb_strlen($message);
         }
 
-        $this->logger->info(sprintf('    reading path %s', $path));
+        $this->write("\r" . '<info>' . str_pad($message, $messageLength, ' ', STR_PAD_RIGHT) . '</info>', false, OutputInterface::VERBOSITY_VERBOSE);
 
-        $finder = new Finder();
-        $finder->files();
-        $finder->name('UserAgentList.php');
-        $finder->ignoreDotFiles(true);
-        $finder->ignoreVCS(true);
-        $finder->sortByName();
-        $finder->ignoreUnreadableDirs();
-        $finder->in($path);
+        $iterator = new RecursiveIteratorIterator(new RecursiveDirectoryIterator(self::PATH));
+        $files    = new class ($iterator, 'php') extends FilterIterator {
+            private string $extension;
 
-        foreach ($finder as $file) {
-            /** @var \Symfony\Component\Finder\SplFileInfo $file */
+            /**
+             * @param Iterator<SplFileInfo> $iterator
+             */
+            public function __construct(Iterator $iterator, string $extension)
+            {
+                parent::__construct($iterator);
+                $this->extension = $extension;
+            }
+
+            public function accept(): bool
+            {
+                $file = $this->getInnerIterator()->current();
+
+                assert($file instanceof SplFileInfo);
+
+                return $file->isFile() && $file->getExtension() === $this->extension;
+            }
+        };
+
+        foreach ($files as $file) {
+            /** @var SplFileInfo $file */
             $filepath = $file->getPathname();
 
-            $this->logger->info('    reading file ' . str_pad($filepath, 100, ' ', STR_PAD_RIGHT));
+            $message = $parentMessage . sprintf('- reading file %s', $filepath);
+
+            if (mb_strlen($message) > $messageLength) {
+                $messageLength = mb_strlen($message);
+            }
+
+            $this->write("\r" . '<info>' . str_pad($message, $messageLength, ' ', STR_PAD_RIGHT) . '</info>', false, OutputInterface::VERBOSITY_VERY_VERBOSE);
 
             $provider = require $filepath;
 
@@ -183,32 +215,35 @@ final class ZsxsoftSource implements SourceInterface
     }
 
     /**
-     * @throws \LogicException
-     * @throws \RuntimeException
-     *
      * @return string[]
+     *
+     * @throws LogicException
+     * @throws RuntimeException
      */
     private function getBrands(): array
     {
         $brands = [];
-        $file   = new \SplFileObject('vendor/zsxsoft/php-useragent/lib/useragent_detect_device.php');
-        $file->setFlags(\SplFileObject::DROP_NEW_LINE);
+        $file   = new SplFileObject('vendor/zsxsoft/php-useragent/lib/useragent_detect_device.php');
+        $file->setFlags(SplFileObject::DROP_NEW_LINE);
         while (!$file->eof()) {
             $line = trim($file->fgets());
             preg_match('/^\$brand = (["\'])(.*)(["\']);$/', $line, $matches);
 
-            if (0 < count($matches)) {
-                $brand = $matches[2];
-                if (!empty($brand)) {
-                    $brands[] = $brand;
-                }
+            if (0 >= count($matches)) {
+                continue;
             }
+
+            $brand = $matches[2];
+            if (empty($brand)) {
+                continue;
+            }
+
+            $brands[] = $brand;
         }
+
         $brands = array_unique($brands);
 
-        usort($brands, static function ($a, $b): int {
-            return mb_strlen($b) - mb_strlen($a);
-        });
+        usort($brands, static fn ($a, $b): int => mb_strlen($b) - mb_strlen($a));
 
         return $brands;
     }
